requirements/base.txt
Django==4.2.7
django-environ==0.11.2
django-localflavor==4.0
django-select2==8.1.2

wagtail==5.2.1

sentry-sdk==1.36.0

psycopg2-binary==2.9.9
django-extensions==3.2.3

xhtml2pdf==0.2.13

gunicorn==21.2.0